feat: allow passing app id + api key directly (#23)
## 🤔 What
- **feat**: allow passing app id + api key directly

## 🤷 Why
It can be practical (for hosting the mcp server) to scope the MCP tools to a specific application instead of relying on the oauth behaviour.
This also allows the node implementation to function in a way that's similar to the [Go implementation](https://github.com/algolia/mcp).

## 🔍 How
Added an optional `--credentials=applicationId:apiKey` cli flags. When present, the MCP server will not use the Dashboard's API tools and endpoint anymore. Instead, the passed credentials will be used directly to query the different public APIs

## 🧪 Testing
Added some ✅